Showcasing capability to adjust ddb's dimension names
This PR simply show cases meshflow's latest version (v0.1.0-dev1)
capability to adjust the ddb's dimension names. This was deemed
necessary as in MESH<r1860 the `gru` dimension could be names `gru`,
however, with MESH>=r1860, the dimension name must be in upper case and
could be named `GRU` or `NGRU`.

The suggestion is `NGRU` to not be mistaken with the `GRU` variable
name.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>